Eureka: extract a bunch of stuff to Types
The main Eureka.hs file was getting crowded. Split out the types.
